+++ arithmetic_lm/model/lightning_module.py
@@ -6,7 +6,7 @@ from arithmetic_lm.eval_utils import eval_on_batch
 from arithmetic_lm.tokenizer import Tokenizer
 from arithmetic_lm.train_utils import lr_cosine_annealing_with_warmup
 
-from .utils import generate
+from .generate import generate
 
 
 class LightningModel(L.LightningModule):
